+++ ingestion/transformers/assign_cadet_domains.py
@@ -1,7 +1,6 @@
-import re
-from typing import Callable, Tuple, Union
+from typing import Callable, Union
 
-import datahub.emitter.mce_builder as builder
+
 from datahub.configuration.common import (KeyValuePattern,
                                           TransformerSemanticsConfigModel)
 from datahub.configuration.import_resolver import pydantic_resolve_key
@@ -9,9 +8,9 @@ from datahub.ingestion.api.common import PipelineContext
 from datahub.ingestion.transformer.dataset_domain import AddDatasetDomain
 from datahub.metadata.schema_classes import DomainsClass
 
-from ingestion.config import ENV, INSTANCE, PLATFORM
-from ingestion.create_derived_table_databases_source.source import \
-    get_cadet_manifest
+from ingestion.dbt_manifest_utils import (
+    get_cadet_manifest, validate_fqn, convert_cadet_manifest_table_to_datahub
+)
 
 
 class AddDatasetDomainSemanticsConfig(TransformerSemanticsConfigModel):
@@ -75,25 +74,10 @@ class AssignDerivedTableDomains(AddDatasetDomain):
             node_info = nodes[node]
             if node_info["resource_type"] != "model":
                 continue
-            domain, escaped_urn_for_regex = convert_cadet_manifest_table_to_datahub(node_info)
-            domain_mappings[escaped_urn_for_regex] = [domain]
+            if validate_fqn(nodes[node]["fqn"]):
+                domain, escaped_urn_for_regex = convert_cadet_manifest_table_to_datahub(node_info)
+                domain_mappings[escaped_urn_for_regex] = [domain]
 
         pattern_input = {"domain_pattern": {"rules": domain_mappings}}
 
         return PatternDatasetDomainSemanticsConfig.parse_obj(pattern_input)
-
-def convert_cadet_manifest_table_to_datahub(node_info: dict) -> Tuple[str, str]:
-    domain = node_info.get("fqn", [])[1]
-    node_table_name = node_info.get("fqn", [])[-1]
-
-    # In CaDeT the convention is to name a table database__table
-    node_table_name_no_double_underscore = node_table_name.replace("__", ".")
-    urn = builder.make_dataset_urn_with_platform_instance(
-            platform=PLATFORM,
-            platform_instance=INSTANCE,
-            env=ENV,
-            name=node_table_name_no_double_underscore,
-        )
-    escaped_urn_for_regex = re.escape(urn)
-
-    return domain, escaped_urn_for_regex
